[REFACTORING] Split IMAPServerIdleSSLTest
